Struct libunbound_sys::ub_server_stats

#[repr(C)]
pub struct ub_server_stats {

    pub num_queries: c_longlong,
    pub num_queries_ip_ratelimited: c_longlong,
    pub num_queries_cookie_valid: c_longlong,
    pub num_queries_cookie_client: c_longlong,
    pub num_queries_cookie_invalid: c_longlong,
    pub num_queries_missed_cache: c_longlong,
    pub num_queries_prefetch: c_longlong,
    pub num_queries_timed_out: c_longlong,
    pub max_query_time_us: c_longlong,
    pub sum_query_list_size: c_longlong,
    pub max_query_list_size: c_longlong,
    pub extended: c_int,
    pub qtype: [c_longlong; 256],
    pub qtype_big: c_longlong,
    pub qclass: [c_longlong; 256],
    pub qclass_big: c_longlong,
    pub qopcode: [c_longlong; 16],
    pub qtcp: c_longlong,
    pub qtcp_outgoing: c_longlong,
    pub qudp_outgoing: c_longlong,
    pub qtls: c_longlong,
    pub qhttps: c_longlong,
    pub qipv6: c_longlong,
    pub qbit_QR: c_longlong,
    pub qbit_AA: c_longlong,
    pub qbit_TC: c_longlong,
    pub qbit_RD: c_longlong,
    pub qbit_RA: c_longlong,
    pub qbit_Z: c_longlong,
    pub qbit_AD: c_longlong,
    pub qbit_CD: c_longlong,
    pub qEDNS: c_longlong,
    pub qEDNS_DO: c_longlong,
    pub ans_rcode: [c_longlong; 16],
    pub ans_rcode_nodata: c_longlong,
    pub ans_secure: c_longlong,
    pub ans_bogus: c_longlong,
    pub rrset_bogus: c_longlong,
    pub queries_ratelimited: c_longlong,
    pub unwanted_replies: c_longlong,
    pub unwanted_queries: c_longlong,
    pub tcp_accept_usage: c_longlong,
    pub ans_expired: c_longlong,
    pub hist: [c_longlong; 40],
    pub msg_cache_count: c_longlong,
    pub rrset_cache_count: c_longlong,
    pub infra_cache_count: c_longlong,
    pub key_cache_count: c_longlong,
    pub msg_cache_max_collisions: c_longlong,
    pub rrset_cache_max_collisions: c_longlong,
    pub num_query_dnscrypt_crypted: c_longlong,
    pub num_query_dnscrypt_cert: c_longlong,
    pub num_query_dnscrypt_cleartext: c_longlong,
    pub num_query_dnscrypt_crypted_malformed: c_longlong,
    pub num_query_dnscrypt_secret_missed_cache: c_longlong,
    pub shared_secret_cache_count: c_longlong,
    pub num_query_dnscrypt_replay: c_longlong,
    pub nonce_cache_count: c_longlong,
    pub num_query_authzone_up: c_longlong,
    pub num_query_authzone_down: c_longlong,
    pub num_neg_cache_noerror: c_longlong,
    pub num_neg_cache_nxdomain: c_longlong,
    pub num_query_subnet: c_longlong,
    pub num_query_subnet_cache: c_longlong,
    pub num_query_cachedb: c_longlong,
    pub mem_stream_wait: c_longlong,
    pub mem_http2_query_buffer: c_longlong,
    pub mem_http2_response_buffer: c_longlong,
    pub qtls_resume: c_longlong,
    pub rpz_action: [c_longlong; 10],

}

per worker statistics.

Fields
num_queries: c_longlong

number of queries from clients received.

num_queries_ip_ratelimited: c_longlong

number of queries that have been dropped/ratelimited by ip.

num_queries_cookie_valid: c_longlong

number of queries with a valid DNS Cookie.

num_queries_cookie_client: c_longlong

number of queries with only the client part of the DNS Cookie.

num_queries_cookie_invalid: c_longlong

number of queries with invalid DNS Cookie.

num_queries_missed_cache: c_longlong

number of queries that had a cache-miss.

num_queries_prefetch: c_longlong

number of prefetch queries - cachehits with prefetch

num_queries_timed_out: c_longlong

number of queries which are too late to process

max_query_time_us: c_longlong

the longest wait time in the queue

sum_query_list_size: c_longlong

Sum of the querylistsize of the worker for every query that missed cache. To calculate average.

max_query_list_size: c_longlong

max value of query list size reached.

extended: c_int

Extended stats below (bool)

qtype: [c_longlong; 256]

qtype stats

qtype_big: c_longlong

bigger qtype values not in array

qclass: [c_longlong; 256]

qclass stats

qclass_big: c_longlong

bigger qclass values not in array

qopcode: [c_longlong; 16]

query opcodes

qtcp: c_longlong

number of queries over TCP

qtcp_outgoing: c_longlong

number of outgoing queries over TCP

qudp_outgoing: c_longlong

number of outgoing queries over UDP

qtls: c_longlong

number of queries over (DNS over) TLS

qhttps: c_longlong

number of queries over (DNS over) HTTPS

qipv6: c_longlong

number of queries over IPv6

qbit_QR: c_longlong

number of queries with QR bit

qbit_AA: c_longlong

number of queries with AA bit

qbit_TC: c_longlong

number of queries with TC bit

qbit_RD: c_longlong

number of queries with RD bit

qbit_RA: c_longlong

number of queries with RA bit

qbit_Z: c_longlong

number of queries with Z bit

qbit_AD: c_longlong

number of queries with AD bit

qbit_CD: c_longlong

number of queries with CD bit

qEDNS: c_longlong

number of queries with EDNS OPT record

qEDNS_DO: c_longlong

number of queries with EDNS with DO flag

ans_rcode: [c_longlong; 16]

answer rcodes

ans_rcode_nodata: c_longlong

answers with pseudo rcode ‘nodata’

ans_secure: c_longlong

answers that were secure (AD)

ans_bogus: c_longlong

answers that were bogus (withheld as SERVFAIL)

rrset_bogus: c_longlong

rrsets marked bogus by validator

queries_ratelimited: c_longlong

number of queries that have been ratelimited by domain recursion.

unwanted_replies: c_longlong

unwanted traffic received on server-facing ports

unwanted_queries: c_longlong

unwanted traffic received on client-facing ports

tcp_accept_usage: c_longlong

usage of tcp accept list

ans_expired: c_longlong

expired answers served from cache

hist: [c_longlong; 40]

histogram data exported to array if the array is the same size, no data is lost, and if all histograms are same size (is so by default) then adding up works well.

msg_cache_count: c_longlong

number of message cache entries

rrset_cache_count: c_longlong

number of rrset cache entries

infra_cache_count: c_longlong

number of infra cache entries

key_cache_count: c_longlong

number of key cache entries

msg_cache_max_collisions: c_longlong

maximum number of collisions in the msg cache

rrset_cache_max_collisions: c_longlong

maximum number of collisions in the rrset cache

num_query_dnscrypt_crypted: c_longlong

number of queries that used dnscrypt

num_query_dnscrypt_cert: c_longlong

number of queries that queried dnscrypt certificates

num_query_dnscrypt_cleartext: c_longlong

number of queries in clear text and not asking for the certificates

num_query_dnscrypt_crypted_malformed: c_longlong

number of malformed encrypted queries

num_query_dnscrypt_secret_missed_cache: c_longlong

number of queries which did not have a shared secret in cache

shared_secret_cache_count: c_longlong

number of dnscrypt shared secret cache entries

num_query_dnscrypt_replay: c_longlong

number of queries which are replays

nonce_cache_count: c_longlong

number of dnscrypt nonces cache entries

num_query_authzone_up: c_longlong

number of queries for unbound’s auth_zones, upstream query

num_query_authzone_down: c_longlong

number of queries for unbound’s auth_zones, downstream answers

num_neg_cache_noerror: c_longlong

number of times neg cache records were used to generate NOERROR responses.

num_neg_cache_nxdomain: c_longlong

number of times neg cache records were used to generate NXDOMAIN responses.

num_query_subnet: c_longlong

number of queries answered from edns-subnet specific data

num_query_subnet_cache: c_longlong

number of queries answered from edns-subnet specific data, and the answer was from the edns-subnet cache.

num_query_cachedb: c_longlong

number of queries served from cachedb

mem_stream_wait: c_longlong

number of bytes in the stream wait buffers

mem_http2_query_buffer: c_longlong

number of bytes in the HTTP2 query buffers

mem_http2_response_buffer: c_longlong

number of bytes in the HTTP2 response buffers

qtls_resume: c_longlong

number of TLS connection resume

rpz_action: [c_longlong; 10]

RPZ action stats